requirements.txt
fastapi
uvicorn
python-decouple
pymongo
itsdangerous
python-multipart





# venv\Scripts\activate